Add store and mapping folders
