+++ lib/datadog/ddsketch/log_collapsing_highest_dense_sketch.rb
@@ -15,9 +15,9 @@ module Datadog
         bin_limit: DEFAULT_BIN_LIMIT
       )
         super(
-          mapping: LogarithmicKeyMapping.new(relative_accuracy: relative_accuracy),
-          store: CollapsingHighestDenseStore.new(bin_limit: bin_limit),
-          negative_store: CollapsingHighestDenseStore.new(bin_limit: bin_limit)
+          mapping: Mapping::LogarithmicKeyMapping.new(relative_accuracy: relative_accuracy),
+          store: Store::CollapsingHighestDenseStore.new(bin_limit: bin_limit),
+          negative_store: Store::CollapsingHighestDenseStore.new(bin_limit: bin_limit)
         )
       end
     end
