+++ lib/datadog/ddsketch/sketch.rb
@@ -10,9 +10,9 @@ module Datadog
     class Sketch < BaseDDSketch
       def initialize(relative_accuracy: DEFAULT_REL_ACC)
         super(
-          mapping: LogarithmicKeyMapping.new(relative_accuracy: relative_accuracy),
-          store: DenseStore.new,
-          negative_store: DenseStore.new
+          mapping: Mapping::LogarithmicKeyMapping.new(relative_accuracy: relative_accuracy),
+          store: Store::DenseStore.new,
+          negative_store: Store::DenseStore.new
         )
       end
     end
