+++ spec/datadog/ddsketch/mapping/cubically_interpolated_key_mapping_spec.rb
@@ -1,4 +1,4 @@
-describe Datadog::DDSketch::CubicallyInterpolatedKeyMapping do
+describe Datadog::DDSketch::Mapping::CubicallyInterpolatedKeyMapping do
   before { skip 'Skipping `Math.ldexp` is inconsistent between JRuby & MRI' if RUBY_PLATFORM == 'java' }
 
   include_context 'mapping tests' do
