+++ spec/datadog/ddsketch/mapping/logarithmic_key_mapping_spec.rb
@@ -1,4 +1,4 @@
-describe Datadog::DDSketch::LogarithmicKeyMapping do
+describe Datadog::DDSketch::Mapping::LogarithmicKeyMapping do
   include_context 'mapping tests' do
     let(:mapping) { described_class.new(relative_accuracy: relative_accuracy, offset: offset) }
   end
